# Improved chat message parsing