--- a/LiveTL/js/lib/filter.js
+++ b/LiveTL/js/lib/filter.js
@@ -1,16 +1,43 @@
+const langTokens = [["[", "]"], ["{", "}"], ["(", ")"], ["|", "|"], ["<", ">"]];
+const startLangTokens = langTokens.flatMap(e => e[0]);
+const tokenMap = Object.fromEntries(langTokens);
+
 /**
  * Parses translation
  *
- * @param msg the message to parse
+ * @param message the message to parse
  * @return undefined or
  * {
  *    lang: lang code
  *    msg: message
  * }
  */
-function parseTranslation (msg) {
-  return surroundFilter(msg) || endFilter(msg)
-}
+const parseTranslation = message => {
+  const trimmed = message.trim();
+
+  const startsWithLeftToken = startLangTokens.includes(trimmed[0]);
+  if (!startsWithLeftToken) {
+    return undefined;
+  }
+
+  const leftToken = trimmed[0];
+  const rightToken = tokenMap[leftToken];
+
+  const righTokenIndex = trimmed.indexOf(rightToken);
+  if (righTokenIndex === -1) {
+    return undefined;
+  }
+
+  // 1 - opening token
+  const lang = trimmed.slice(1, righTokenIndex);
+  // 1 - closing token, trim - may be space after [lang] block
+  const msg = trimmed.slice(righTokenIndex + 1).trim();
+
+  return {
+    lang,
+    msg,
+  };
+};
 
 function isLangMatch (textLang, currentLang) {
   textLang = textLang.toLowerCase().split(/[\/\ \-\:\.\|]/).filter(s => s !== '')
@@ -20,58 +47,3 @@ function isLangMatch (textLang, currentLang) {
     currentLang.lang.toLowerCase().startsWith(s)
   ))
 }
-
-function createSurroundRegex () {
-  const surroundTokens = [
-    '()', '[]', '{}', '||', '<>'
-  ]
-  let pattern = ''
-  let patternEnd = ''
-  let notPattern = ''
-
-  surroundTokens.forEach((token) => {
-    pattern = `${pattern}\\${token[0]}`
-    patternEnd = `${patternEnd}\\${token[1]}`
-    notPattern = `${notPattern}^\\${token[0]}^\\${token[1]}`
-  })
-
-  return new RegExp(
-    `^([${pattern}])([${notPattern}]+)([${patternEnd}]) ?[\-\:\.\|]? ?(.+)`
-  )
-}
-
-function surroundTokensMatch (token1, token2) {
-  switch (token1) {
-    case '(':
-      return token2 === ')'
-    case '[':
-      return token2 === ']'
-    case '{':
-      return token2 === '}'
-  }
-  return token1 === token2
-}
-
-function surroundFilter (msg) {
-  const surroundRegex = createSurroundRegex()
-  const result = surroundRegex.exec(msg)
-  if (result && surroundTokensMatch(result[1], result[3])) {
-    return {
-      lang: result[2].trim(),
-      msg: result[4],
-    }
-  }
-}
-
-function endFilter (msg) {
-  const result = /^([^\-^\:^\|]+)[\-\:\.\|] ?(.+)/.exec(msg)
-  if (result) {
-    return {
-      lang: result[1].trim(),
-      msg: result[2],
-    }
-  }
-}
-
-module.exports = { parseTranslation, isLangMatch }
-