--- a/LiveTL/js/lib/filter.js
+++ b/LiveTL/js/lib/filter.js
@@ -47,3 +47,5 @@ function isLangMatch (textLang, currentLang) {
     currentLang.lang.toLowerCase().startsWith(s)
   ))
 }
+
+module.exports = { parseTranslation, isLangMatch };