@Dragon1320 2020-12-03
ahh ye didnt realise there were other ways of denoting translations other than brackets, gonna fix that now