@r2dev2bb8 2020-12-03
Initial benchmarks don't show improved speed, I'm writing better benchmarks which simulate the scenarios you provided to get a better reading of the performance .